backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "1.0.0"
description = ""
authors = ["Kirill Zhosul <kirillzhosul@yandex.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.4"
djangorestframework = "^3.15.1"
channels = { extras = ["daphne"], version = "^4.1.0" }
channels-redis = "^4.2.0"
djoser = "^2.2.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"

[tool.poetry.group.dev.dependencies]
inflection = "^0.5.1"
uritemplate = "^4.1.1"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
